remove first require cycles
