+++ src/libs/addTrailingForwardSlash.ts
@@ -1,0 +1,11 @@
+/**
+ * Add / to the end of any URL if not present
+ */
+function addTrailingForwardSlash(url: string): string {
+    if (!url.endsWith('/')) {
+        return `${url}/`;
+    }
+    return url;
+}
+
+export default addTrailingForwardSlash;
