+++ src/libs/models/BankAccountState.ts
@@ -1,0 +1,10 @@
+const BankAccountState = {
+    PENDING: 'PENDING',
+    OPEN: 'OPEN',
+    DELETED: 'DELETED',
+    LOCKED: 'LOCKED',
+    SETUP: 'SETUP',
+    VERIFYING: 'VERIFYING',
+};
+
+export default BankAccountState;
